Toggle Physics Option
Added an option in the settings tab labeled "toggle physics" which will
disable and enable physics.

I wrote this code a week ago and was attempting to debug it but was
unable to pull up the dev-demo because my branch was too far out of sync
with the master branch.  When I attempted to sync up with the master
branch I was getting compiler errors (possibly caused by me having
Qt5.5). Tom has encouraged me to submit my code despite not being able
to compile or test it for review so please let me know if it has a bug
or if there is anything you would like me to add.

The problem that I chose to solve with this submission is the creation
of a drop down tab that would disable physics until the user reenabled
it again. I accomplished this by adding in a flag inside of Application
(interface\src\Application.h) called _physicsManuallyDisabled which
would store if the user disabled or enabled physics from the drop menu.
I added in the flag inside of the update() when the _physicsEnabled flag
is checked so that if physics is manually disabled then _physicsEnabled
will never be allowed to be set high until physics is enabled again.
Lastly, I control the value of _physicsManuallyDisabled through an
addition to the settings dropdown menu with I labeled "toggle physics".
When the user selects this option it calls the togglePhysics() function
which toggles the _physicsManuallyDisabled flag (as well as sets
_physicsEnabled low if physics is disabled).

Please feel free to contact me if you are having any issues with my code
or would like for me to provide any more clarity.

Ryan J